+++ opal/legacy/api/views.py
@@ -1,17 +1,17 @@
 """Collection of api views used to send data to opal app through the listener request relay."""
 
-from django.db.models import QuerySet
-from django.http import HttpRequest, HttpResponse
-from django.utils import timezone
+from typing import Any
 
 from rest_framework.permissions import IsAuthenticated
+from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from opal.legacy.utils import get_patient_sernum
+from opal.core.drf_permissions import CaregiverPatientPermissions
 
-from ..models import LegacyAppointment, LegacyNotification
-from .serializers import LegacyAppointmentSerializer
+from .. import models
+from ..utils import get_patient_sernum
+from .serializers import LegacyAppointmentSerializer, UnreadCountSerializer
 
 
 class AppHomeView(APIView):
@@ -19,7 +19,7 @@ class AppHomeView(APIView):
 
     permission_classes = [IsAuthenticated]
 
-    def get(self, request: HttpRequest) -> HttpResponse:
+    def get(self, request: Request) -> Response:
         """
         Handle GET requests from `api/app/home`.
 
@@ -31,43 +31,73 @@ class AppHomeView(APIView):
         """
         patient_sernum = get_patient_sernum(request.headers['Appuserid'])
         return Response({
-            'unread_notification_count': self.get_unread_notification_count(patient_sernum),
+            'unread_notification_count': models.LegacyNotification.objects.get_unread_queryset(patient_sernum).count(),
             'daily_appointments': LegacyAppointmentSerializer(
-                self.get_daily_appointments(patient_sernum),
+                models.LegacyAppointment.objects.get_daily_appointments(patient_sernum),
                 many=True,
             ).data,
         })
 
-    def get_daily_appointments(self, patient_sernum: int) -> QuerySet[LegacyAppointment]:
+
+class AppChartView(APIView):
+    """Class to return chart page required data."""
+
+    permission_classes = [IsAuthenticated]
+
+    def get(self, *args: Any, **kwargs: Any) -> Response:
         """
-        Get all appointment for the current day.
+        Handle GET requests from `api/app/chart`.
+
+        The function provides the number of unread values for the user
+        and will provide them for the selected patient instead until the profile selector is finished.
 
         Args:
-            patient_sernum: Patient sernum used to retrieve unread notifications count.
+            args: additional arguments
+            kwargs: additional keyword arguments
 
         Returns:
-            Appointments schedule for the current day.
+            Http response with the data needed to display the chart view.
         """
-        return LegacyAppointment.objects.select_related(
-            'aliasexpressionsernum',
-            'aliasexpressionsernum__aliassernum',
-            'aliasexpressionsernum__aliassernum__appointmentcheckin',
-        ).filter(
-            scheduledstarttime__date=timezone.localtime(timezone.now()).date(),
-            patientsernum=patient_sernum,
-            state='Active',
-        ).exclude(
-            status='Deleted',
-        )
+        legacy_id = kwargs['legacy_id']
+        unread_count = {
+            'unread_appointment_count': models.LegacyAppointment.objects.get_unread_queryset(legacy_id).count(),
+            'unread_document_count': models.LegacyDocument.objects.get_unread_queryset(legacy_id).count(),
+            'unread_txteammessage_count': models.LegacyTxTeamMessage.objects.get_unread_queryset(
+                legacy_id,
+            ).count(),
+            'unread_educationalmaterial_count': models.LegacyEducationalMaterial.objects.get_unread_queryset(
+                legacy_id,
+            ).count(),
+            'unread_questionnaire_count': models.LegacyQuestionnaire.objects.get_unread_queryset(
+                legacy_id,
+            ).count(),
+        }
 
-    def get_unread_notification_count(self, sernum: int) -> int:
+        return Response(UnreadCountSerializer(unread_count).data)
+
+
+class CaregiverPermissionsView(APIView):
+    """
+    Bare-bones API view used to perform a simple caregiver-patient permissions check.
+
+    Validates that a caregiver has permission to view data for a specific patient (based on their relationship).
+    Used by the legacy part of the listener to access functionality from CaregiverPatientPermissions.
+    """
+
+    # The essential work for this request is done by CaregiverPatientPermissions
+    permission_classes = [IsAuthenticated, CaregiverPatientPermissions]
+
+    def get(self, request: Request, legacy_id: int) -> Response:
         """
-        Get the number of unread notifications for a given user.
+        Handle GET requests on `patients/legacy/<legacy_id>/check_permissions`.
 
         Args:
-            sernum: User sernum used to retrieve unread notifications count.
+            request: Http request made by the listener.
+            legacy_id: The legacy ID used to represent the target patient.
 
         Returns:
-            Number of unread notifications.
+            Http response containing no data if the permission checks succeed (200 OK), or a permission denied
+            response if the permission checks fail.
         """
-        return LegacyNotification.objects.filter(patientsernum=sernum, readstatus=0).count()
+        # The following empty success response is only returned if CaregiverPatientPermissions succeeds
+        return Response({})