+++ CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Feat: Get all data needed for the app home view request (!118)
 * feat: Registration encryption info endpoint (!170)
 * Add logo (image) field to the Institution model (!159)
+* feat: create permissions class to check caregiver-patient access (!169)
+* Adjust legacy chart api endpoint to use legacy_id (!185)
+* feat: add model for a user's mobile device (!187)
+* Add hospital service for communication with the OIE (!173)
 
 ### Documentation